feat: products page, home page and new title
