+++ src/types/product.ts
@@ -1,5 +1,7 @@
 export interface Product {
   id: number
+  brand: string
+  model: string
   title: string
   image: string
   price: number
